// src/types/eventTypes.ts
// types.ts

export enum EventCategory {
  MUSIC = "MUSIC",
  LIFESTYLE_EVENTS = "LIFESTYLE_EVENTS",
  HOLIDAY_CELEBRATIONS = "HOLIDAY_CELEBRATIONS",
  BUSINESS = "BUSINESS",
  CONCERTS_AND_CHURCH = "CONCERTS_AND_CHURCH",
  CONFERENCES_AND_WORKSHOPS = "CONFERENCES_AND_WORKSHOPS",
  SPORTS_AND_FITNESS = "SPORTS_AND_FITNESS",
  WEB_DEVELOPMENT = "WEB_DEVELOPMENT",
  MARKETING = "MARKETING",
  TECHNOLOGY = "TECHNOLOGY",
  EDUCATIONAL_EVENTS = "EDUCATIONAL_EVENTS",
  FASHION_SHOWS = "FASHION_SHOWS",
  HEALTH_AND_WELLNESS = "HEALTH_AND_WELLNESS",
  CULTURAL_FESTIVALS = "CULTURAL_FESTIVALS",
  GAMING_EVENTS = "GAMING_EVENTS",
  ENVIRONMENTAL_EVENTS = "ENVIRONMENTAL_EVENTS",
  TRADE_FAIR = "TRADE_FAIR",
  WEDDING = "WEDDING",
  FUNERAL = "FUNERAL",
  AGRICULTURAL_AND_COMMERCIAL_SHOW = "AGRICULTURAL_AND_COMMERCIAL_SHOW",
  ARTS_AND_THEATER = "ARTS_AND_THEATER",
  FAMILY_AND_KIDS = "FAMILY_AND_KIDS",
  FOOD_AND_DRINK = "FOOD_AND_DRINK",
  CHARITY_AND_FUNDRAISERS = "CHARITY_AND_FUNDRAISERS",
  NETWORKING_AND_SOCIAL_GATHERINGS = "NETWORKING_AND_SOCIAL_GATHERINGS",
  FILM_SCREENINGS = "FILM_SCREENINGS",
}

export enum EventType {
  PHYSICAL = "PHYSICAL",
  ONLINE = "ONLINE",
}

export interface EventFormData {
  image: File | null;
  imagePreview: string | null;
  title: string;
  description: string;
  dateStart: string;
  dateEnd: string;
  location: string;
  category: EventCategory;
  timeStart: string;
  timeEnd: string;
  totalSeats: number;
  vipSeats: number;
  regularPrice: number;
  vipPrice: number;
  isFree: boolean;
  eventType: EventType;
}